Refactor: (BRD-89) 리다이렉트 필터 설정을 FilterChainConfig 에서 함께 설정하도록 수정
